Move parser under utils
