+++ services/blockchain-connector/shared/utils/parser.js
